emecAPI/utils/fields.py
from unicodedata    import normalize

def normalize_key(key: str) -> str:
    """
    Normalize a given key by removing leading and trailing whitespace and colons,
    converting it to lowercase, replacing spaces with underscores, removing parentheses,
    removing non-ascii characters, and decoding the string.

    Args:
        key (str): The key to be normalized.

    Returns:
        str: The normalized key.
    """
    text = key.strip(': ')                              # Remove leading and trailing whitespace and colons.
    text = text.lower()                                 # Convert to lowercase.
    text = text.replace(' ', '_')                       # Replace spaces with underscores.
    text = text.replace('(' , '').replace(')', '')      # Remove parentheses.

    normalized = normalize('NFKD', text)                # Normalize the string.
    normalized = normalized.encode('ascii', 'ignore')   # Remove non-ascii characters.
    normalized = normalized.decode('utf-8')             # Decode the string.

    return normalized